Fix propagation of service accounts for site replication
- Only non-root owned service accounts are replicated for now.

- Add integration tests for OIDC with site replication
